Refactored current adapter code into `rapporto.github`
... to accompany including other adapters.
